Replace Option with Optional
Replaces org.opencastproject.util.data.Option with java.util.Optional.
The goal here is to avoid custom reimplementation of standard Java functionality.
There should be no functional changes.

+++ modules/workspace-impl/src/main/java/org/opencastproject/workspace/impl/WorkspaceImpl.java
@@ -31,8 +31,6 @@ import static org.opencastproject.util.RequireUtil.notNull;
 import static org.opencastproject.util.data.Arrays.cons;
 import static org.opencastproject.util.data.Either.left;
 import static org.opencastproject.util.data.Either.right;
-import static org.opencastproject.util.data.Option.none;
-import static org.opencastproject.util.data.Option.some;
 import static org.opencastproject.util.data.Prelude.sleep;
 
 import org.opencastproject.assetmanager.util.AssetPathUtils;
@@ -50,7 +48,6 @@ import org.opencastproject.util.PathSupport;
 import org.opencastproject.util.data.Effect;
 import org.opencastproject.util.data.Either;
 import org.opencastproject.util.data.Function;
-import org.opencastproject.util.data.Option;
 import org.opencastproject.util.data.functions.Misc;
 import org.opencastproject.util.jmx.JmxUtil;
 import org.opencastproject.workingfilerepository.api.PathMappable;
@@ -91,6 +88,7 @@ import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CopyOnWriteArraySet;
 
@@ -463,26 +461,26 @@ public final class WorkspaceImpl implements Workspace {
    * @throws IOException
    *           in case of any IO related issues
    */
-  private Either<String, Option<File>> handleDownloadResponse(HttpResponse response, URI src, File dst)
+  private Either<String, Optional<File>> handleDownloadResponse(HttpResponse response, URI src, File dst)
           throws IOException {
     final String url = src.toString();
     final int status = response.getStatusLine().getStatusCode();
     switch (status) {
       case HttpServletResponse.SC_NOT_FOUND:
-        return right(none(File.class));
+        return right(Optional.empty());
       case HttpServletResponse.SC_NOT_MODIFIED:
         logger.debug("{} has not been modified.", url);
-        return right(some(dst));
+        return right(Optional.of(dst));
       case HttpServletResponse.SC_ACCEPTED:
         logger.debug("{} is not ready, try again later.", url);
         return left(response.getHeaders("token")[0].getValue());
       case HttpServletResponse.SC_OK:
         logger.debug("Downloading {} to {}", url, dst.getAbsolutePath());
-        return right(some(downloadTo(response, dst)));
+        return right(Optional.of(downloadTo(response, dst)));
       default:
         logger.warn("Received unexpected response status {} while trying to download from {}", status, url);
         FileUtils.deleteQuietly(dst);
-        return right(none(File.class));
+        return right(Optional.empty());
     }
   }
 
@@ -516,7 +514,7 @@ public final class WorkspaceImpl implements Workspace {
       // run the http request and handle its response
       try {
         HttpResponse response = null;
-        final Either<String, Option<File>> result;
+        final Either<String, Optional<File>> result;
         try {
           response = trustedHttpClient.execute(get);
           result = handleDownloadResponse(response, src, dst);
@@ -525,9 +523,9 @@ public final class WorkspaceImpl implements Workspace {
             trustedHttpClient.close(response);
           }
         }
-        for (Option<File> ff : result.right()) {
-          for (File f : ff) {
-            return f;
+        for (Optional<File> ff : result.right()) {
+          if (ff.isPresent()) {
+            return ff.get();
           }
           FileUtils.deleteQuietly(dst);
           // none
@@ -873,18 +871,18 @@ public final class WorkspaceImpl implements Workspace {
   }
 
   @Override
-  public Option<Long> getTotalSpace() {
-    return some(new File(wsRoot).getTotalSpace());
+  public Optional<Long> getTotalSpace() {
+    return Optional.of(new File(wsRoot).getTotalSpace());
   }
 
   @Override
-  public Option<Long> getUsableSpace() {
-    return some(new File(wsRoot).getUsableSpace());
+  public Optional<Long> getUsableSpace() {
+    return Optional.of(new File(wsRoot).getUsableSpace());
   }
 
   @Override
-  public Option<Long> getUsedSpace() {
-    return some(FileUtils.sizeOfDirectory(new File(wsRoot)));
+  public Optional<Long> getUsedSpace() {
+    return Optional.of(FileUtils.sizeOfDirectory(new File(wsRoot)));
   }
 
   @Override
